[功能重构] 优化房间代码生成逻辑，改用Base36编码并添加前缀支持，增强安全性和可读性

+++ MVL/Src/Utils/Multiplayer/RoomPlayerInfo.cs
@@ -8,7 +8,7 @@ namespace MVL.Utils.Multiplayer;
 public partial record RoomPlayerInfo(
 	RoomType RoomType,
 	string Name,
-	int Port,
+	ushort Port,
 	string Address,
 	string Version) {
 	public uint Identity { get; set; }
